Avoid integer overflow and sort out casts.
Increase the size of BlockStatusData.Length to an int64, to avoid
overflow in GetBlockStatus. Also change BlockStatusData.Offset to an
int64 and remove a handful of unnerving integer conversions. Favor int64
over uint64 to match system libraries, and add a few necessary
conversions mostly isolated to libnbd interfaces.

+++ pkg/importer/vddk-datasource_test.go
@@ -460,10 +460,10 @@ var _ = Describe("VDDK data source", func() {
 		}
 		_, err := NewVDDKDataSource("http://esx.test", "user", "pass", "aa:bb:cc:dd", "1-2-3-4", diskName, "", "", "", v1.PersistentVolumeFilesystem)
 		Expect(err).ToNot(HaveOccurred())
-		Expect(MaxPreadLength).To(Equal(uint32(MaxPreadLengthESX)))
+		Expect(MaxPreadLength).To(Equal(MaxPreadLengthESX))
 		_, err = NewVDDKDataSource("http://vcenter.test", "user", "pass", "aa:bb:cc:dd", "1-2-3-4", diskName, "", "", "", v1.PersistentVolumeFilesystem)
 		Expect(err).ToNot(HaveOccurred())
-		Expect(MaxPreadLength).To(Equal(uint32(MaxPreadLengthVC)))
+		Expect(MaxPreadLength).To(Equal(MaxPreadLengthVC))
 	})
 })
 
@@ -531,8 +531,8 @@ var _ = Describe("VDDK get block status", func() {
 		}
 		expectedBlocks := []*BlockStatusData{
 			{
-				Offset: uint64(extent.Start),
-				Length: uint32(extent.Length),
+				Offset: extent.Start,
+				Length: extent.Length,
 				Flags:  0,
 			},
 		}
@@ -555,8 +555,8 @@ var _ = Describe("VDDK get block status", func() {
 		}
 		expectedBlocks := []*BlockStatusData{
 			{
-				Offset: uint64(extent.Start),
-				Length: uint32(extent.Length),
+				Offset: extent.Start,
+				Length: extent.Length,
 				Flags:  0,
 			},
 		}
@@ -577,8 +577,8 @@ var _ = Describe("VDDK get block status", func() {
 		}
 		expectedBlocks := []*BlockStatusData{
 			{
-				Offset: uint64(extent.Start),
-				Length: uint32(extent.Length),
+				Offset: extent.Start,
+				Length: extent.Length,
 				Flags:  0,
 			},
 		}
@@ -610,8 +610,8 @@ var _ = Describe("VDDK get block status", func() {
 		}
 		expectedBlocks := []*BlockStatusData{
 			{
-				Offset: uint64(extent.Start),
-				Length: uint32(extent.Length),
+				Offset: extent.Start,
+				Length: extent.Length,
 				Flags:  0,
 			},
 		}
@@ -665,8 +665,8 @@ var _ = Describe("VDDK get block status", func() {
 		}
 		expectedBlocks := []*BlockStatusData{
 			{
-				Offset: uint64(extent.Start),
-				Length: uint32(extent.Length),
+				Offset: extent.Start,
+				Length: extent.Length,
 				Flags:  0,
 			},
 		}
@@ -703,8 +703,8 @@ var _ = Describe("VDDK get block status", func() {
 
 		blocks := GetBlockStatus(&mockNbdOperations{}, extent)
 		for index, block := range blocks {
-			Expect(block.Offset).To(Equal(uint64(index * MaxBlockStatusLength)))
-			Expect(block.Length).To(Equal(uint32(MaxBlockStatusLength)))
+			Expect(block.Offset).To(Equal(int64(index * MaxBlockStatusLength)))
+			Expect(block.Length).To(Equal(int64(MaxBlockStatusLength)))
 			if index%2 == 0 {
 				Expect(block.Flags).To(Equal(uint32(0)))
 			} else {
@@ -795,9 +795,9 @@ type mockVddkDataSink struct {
 	position int
 }
 
-func (sink *mockVddkDataSink) ZeroRange(offset uint64, length uint32) error {
+func (sink *mockVddkDataSink) ZeroRange(offset int64, length int64) error {
 	buf := bytes.Repeat([]byte{0x00}, int(length))
-	_, err := sink.Pwrite(buf, offset)
+	_, err := sink.Pwrite(buf, uint64(offset))
 	return err
 }
